+++ gui/src/helpers/__init__.py
@@ -7,6 +7,7 @@ from .core.duplicate_scan_worker import DuplicateScanWorker as DuplicateScanWork
 from .core.merge_worker import MergeWorker as MergeWorker
 from .core.sampler_worker import SamplerWorker as SamplerWorker
 from .core.similarity_scan_worker import SimilarityScanWorker as SimilarityScanWorker
+from .database.embedding_worker import ImageEmbeddingWorker as ImageEmbeddingWorker
 from .database.search_worker import SearchWorker as SearchWorker
 from .database.upsert_worker import UpsertWorker as UpsertWorker
 from .core.video_export_worker import ScrollVideoExportWorker as ScrollVideoExportWorker